Crash with Motorola Droid
The stacktrace hints together with the source code and information from the internet that a primary color with transparency is invalid. I couldn't reproduce the bug but all the information leads to the conclusion that this is the error. Unfortunately, it is not clear why this doesn't error doesn't show up every time with every device.

For a fast fix, deeper analysis is postponed.